updated note & generator hierarchy

+++ src/generator/saw_tooth.rs
@@ -2,12 +2,12 @@ use crate::generator::ToneGenerator;
 
 #[derive(Debug)]
 pub struct SawTooth {
-    period: f64,
-    amplitude: f64,
+    period: f32,
+    amplitude: f32,
 }
 
 impl SawTooth {
-    pub fn new(frequency: f64, amplitude: f64) -> Self {
+    pub fn new(frequency: f32, amplitude: f32) -> Self {
         Self {
             period: 1.0 / frequency,
             amplitude,
@@ -16,7 +16,7 @@ impl SawTooth {
 }
 
 impl ToneGenerator for SawTooth {
-    fn generate(&self, time: f64) -> f64 {
+    fn generate(&self, time: f32) -> f32 {
         self.amplitude * (
             2.0 * (( time / self.period ) - ((1.0 / 2.0) + (time / self.period)).floor())
         )
