+++ src/generator/sine_wave.rs
@@ -1,14 +1,14 @@
 use crate::generator::ToneGenerator;
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 
 #[derive(Debug)]
 pub struct SineWave {
-    frequency: f64,
-    amplitude: f64,
+    frequency: f32,
+    amplitude: f32,
 }
 
 impl SineWave {
-    pub fn new(frequency: f64, amplitude: f64) -> Self {
+    pub fn new(frequency: f32, amplitude: f32) -> Self {
         Self {
             frequency,
             amplitude,
@@ -17,7 +17,7 @@ impl SineWave {
 }
 
 impl ToneGenerator for SineWave {
-    fn generate(&self, time: f64) -> f64 {
+    fn generate(&self, time: f32) -> f32 {
         self.amplitude * (2.0 * PI * self.frequency * time).sin()
     }
 }
