+++ src/generator/square_wave.rs
@@ -2,12 +2,12 @@ use crate::generator::ToneGenerator;
 
 #[derive(Debug)]
 pub struct SquareWave {
-    frequency: f64,
-    amplitude: f64,
+    frequency: f32,
+    amplitude: f32,
 }
 
 impl SquareWave {
-    pub fn new(frequency: f64, amplitude: f64) -> Self {
+    pub fn new(frequency: f32, amplitude: f32) -> Self {
         Self {
             frequency,
             amplitude,
@@ -16,7 +16,7 @@ impl SquareWave {
 }
 
 impl ToneGenerator for SquareWave {
-    fn generate(&self, time: f64) -> f64 {
+    fn generate(&self, time: f32) -> f32 {
         self.amplitude * (
             2.0 * ( 2.0 * ( time * self.frequency ).floor() - (2.0 * time * self.frequency).floor()) + 1.0
         )
